Update AbstractGraph.removeAllEdges(V, V) to match specification.
The documentation for Graph.removeAllEdges(V, V) specifies that if any of the vertices are not in the graph, the method should return null. The previous implementation throws a NullPointerException instead, since getAllEdges() is also specified to return null if one of the vertices doesn't exist, and removeAllEdges(Collection<? extends E>) throws NullPointerException when called on a null argument. The current change fixes this.