IntrusiveEdgesSpecifics: reduce Map lookups and prevent invalid edge reuse (#1034)
* [IntrusiveEdgesSpecifics] reduce Map lookups and check for edge reuse

- Reduce number of map lookups from two to one when adding new edges.

- Detect if an IntrusiveEdge was already used in another graph with
different touching vertices and reject the edge in this case (do not
overwrite its source and target).

* [EdgesSpecifics] throw exception if edge already connects other vertices

also unify adding of intrusive edges and handling of rejected edges.

* Remove up-front check in addEdge(V, V, E) if edge is already contained

In most cases this check is just a performance penalty. And because the
rejection of an edge from the EdgeSpecific has to be handled anyway
since it can throw an exception there is no need to ensure up front that
adding will succeed.
An up-front check only saves runtime in case the edge is already present
in this or another graph but with other touching vertices (and is not an
intrusive edge). Compared to other cases (for example if just a valid
edge is added) it is probably rare. Also if it can be decided which case
to prefer it should be the valid one.

* Implement remarks of @d-michail

* Make IntrusiveEdgeException a public top-level class

* Add copy-right header and doc and extend RuntimeException