Improvements of StrongConnectivityInspector algorithms (#1020)
* [SCC] Improvements of StrongConnectivityInspector-algorithms

GabowStrongConnectivityInspector:
- replace Vector by ArrayList
- create HashSets holding the SCC's vertices with the size of each SCC
and use Collections.singleton() for single vertex SCCs

- cosmetic changes:
  - simplify method signatures (just use graph field)
  - rename field "stack" to "S" to comply to paper of Gabow
  - use stack-methods (push()/pop()/peek()) instead of Dequeue methods
  - inline methods of VertexNumber (also could reduce Object creation)

KosarajuStrongConnectivityInspector:
- replace Vector by ArrayList

AbstractStrongConnectivityInspector:
- use edge-supplier
- create HashMap with expected size

StrongConnectivityAlgorithmTest:
- use parameterized JUnit runner to test different implementations
- simplify/unify graph creation and assertions/checks of computed SCCs
- add more test cases

* [Gabow-SCC] avoid Integer-objects and left behind initial node

- use one-based vertex numbers instead of zero-based. Using zero-based
numbers caused the first node of a DFS to be on the stacks twice and
also to remain on stack S and B after DFS has completed, if the first
node is part of an SCC. This also had the effect that cycles that
contain the initial node are detected not before a successor of the
initial node was explored. The example of a three vertex directed ring,
makes this issue clear.

- using one-based numbers allows to use VertexNumber objects in stack B
too. This avoids creation of Integer-objects (which are only cached up
to 127 by default. See Integer$IntegerCache). This is possible because
each vertex is on stack S (and B) at most once and has a unique number
(as long as it is on S).

* [GabowStrongConnectivityInspector] prefix stack-fields with "stack"

AbstractStrongConnectivityInspector use the class-argument constructor
again. The SupplierUtil now uses a performant and a proper
(serializable) supplier.

* Create HashSet with expected size and make constructor protected